refactor!: use static variables/methods over ctr parameters
The status bar and the window title prefix are meant to be global for
all open projects. Hence, there is no reason for this to be passed in on
object creation and also is in conflict with responsibility driven
design. I have moved to providing static methods in because it seems
more sensible in the long run.